\name{list_output_paths}
\alias{list_output_paths}
\title{MadingleyR: Get the output file paths}
\usage{list_output_paths(madingley_data, full_path = TRUE)}
\description{This functions returns the files names and/or paths of the files written by the MadingleyR package. It only works when an output folder is specified in the madingley_run() function. This aids the reading of files for processing at a later stage (see example below).}

\arguments{
 \item{madingley_data}{output generated by madingley_run()}
 \item{full_path}{whether the function should return the full path (full_path = TRUE) or the names of the files only (full_path = FALSE)}
}

\examples{
# Set output folder
out_dir = 'path/to/outputs'

# Set spatial window
spatial_window = c(31, 35, -5, -1)

# Call the madingley_init function (default run)
m_data = madingley_init(spatial_window = spatial_window)

# Run the model
m_data2 = madingley_run(out_dir = out_dir,
                        madingley_data = m_data,
                        years = 4)

# Get the full paths of the output files
output_paths = list_output_paths(madingley_data = m_data2)

# Get names of the output files
output_names = list_output_paths(madingley_data = m_data2, full_path = F)

# Use output_paths to load a file written
cat(output_paths$cohort_properties[1])
BinnedCohortStatistics1 = read.csv(output_paths$cohort_properties[1])
head(BinnedCohortStatistics1)
}
