\name{madingley_plot}
\alias{madingley_plot}
\title{MadingleyR: Create MadingleyR plots}
\usage{madingley_plot(madingley_data)}

\description{
Create all output plots included in the MadinleyR package. This function is intended as a tool to quickly visualize the results.
All resulting plots can also be made individually by using the following functions:

- plot_densities()

- plot_foodweb()

- plot_spatialabundances()

- plot_spatialbiomass()

- plot_timelines()

- plot_trophicpyramid()

}

\arguments{\item{madingley_data}{output generated by madingley_run()}}

\examples{
# Print all the available options to load to the R console
madingley_inputs( )

# Load spatial parameters
sptl_inp = madingley_inputs("spatial inputs")

# Load default cohort property definitions
chrt_def = madingley_inputs("cohort definition")

# Load default stock property definitions
stck_def = madingley_inputs("stock definition")

# Load default model parameters
mdl_prms = madingley_inputs("model parameters")

# Set spatial model domain
spatial_window = c(31, 35, -5, -1)

# Call the madingley_init function
m_data = madingley_init(cohort_def = chrt_def,
                        stock_def = stck_def,
                        spatial_inputs = sptl_inp,
                        spatial_window = spatial_window,
                        max_cohort = 500)

# Run the madingley model using the m_data as input
m_data2 = madingley_run(madingley_data = m_data,
                        years = 5,
                        max_cohort = 500,
                        cohort_def = chrt_def,
                        stock_def = stck_def,
                        spatial_inputs = sptl_inp,
                        model_parameters = mdl_prms)

# Create the MadingleyR plots
madingley_plot(m_data2)
}
