\name{madingley_version}
\alias{madingley_version}
\title{MadingleyR: Get version of MadingleyR and C++ source code}
\usage{
# Get version MadingleyR and C++ source code
mdata = madingley_version()
}
\description{This function provides the version of the MadingleyR package as well as the version of the Madingley C++ source code.}
